requirements.txt
pillow
timm